feat: begin last.fm integration
This adds a non-functional LastfmCard.tsx component and LastfmService.ts service for use in the upcoming Last.fm integration.

The idea currently is to send a request from the client to the backend, which in turn will send a request to the last.fm api, which will return the information to the backend and to the frontend. Doing it this way willa void exposing API keys in the frontend.