style: remove background image and resize lastfm card
This commit removes the background image, as it proved too problematic to make look good across desktop and mobile browser windows. Added a soft grey gradient as the background for the time being. Also set the max width of the lastfm card so that it is only 50% of the size of full width cards.