fix: main container stretching to fill page
This commit fixes the issue where the container would stretch vertically to fill 95vh when the content inside was not enough to justify the length. This commit also adds functionality so that the container does limit itself to 95vh when the height of the window is less than the height of the container. However, the current method of doing this is very manual, and will need updating when more content is added to the main panel.